feat: categories for cloudwatch log entries
Support the ability for a cloud watch log event to explicitly include a
unique list of categories that it should be saved under.

If no explicit categories are provided, the event is always considered
to carry the "Logs" category.

If multiple categories, return a record per category.

Provide the category as a partition value.